Data flow between scenes for start screen and main game.
Adds an example start screen scene that send you to SampleScene.

Press x and/or p on the start screen for player 1 and/or 2 to join
the game, then press enter to start the game. Should work for controllers
as well, but I haven't tested it. Players 3 and 4 should work, but can
only join the game using controllers (button 0) and I haven't tested that
either.